+++ src/test/java/com/airbus_cyber_security/graylog/events/processor/aggregation/checks/AggregationCountTest.java
@@ -21,12 +21,17 @@
 package com.airbus_cyber_security.graylog.events.processor.aggregation.checks;
 
 import com.airbus_cyber_security.graylog.events.processor.aggregation.AggregationCountProcessorConfig;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
+import org.graylog.events.notifications.EventNotificationSettings;
+import org.graylog.events.processor.EventDefinition;
+import org.graylog.events.processor.EventDefinitionDto;
+import org.graylog.events.processor.aggregation.*;
 import org.graylog.events.search.MoreSearch;
 import org.graylog2.indexer.results.CountResult;
 import org.graylog2.indexer.results.ResultMessage;
 import org.graylog2.indexer.results.SearchResult;
-import org.graylog2.indexer.results.TermsResult;
+import org.graylog2.indexer.searches.Searches;
 import org.graylog2.indexer.searches.Sorting;
 import org.graylog2.plugin.indexer.searches.timeranges.AbsoluteRange;
 import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
@@ -40,25 +45,57 @@ import org.mockito.junit.MockitoRule;
 
 import java.util.*;
 
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.hamcrest.CoreMatchers.containsString;
 
 
 public class AggregationCountTest {
 
-    private final int threshold = 100;
+    private final int threshold = 2;
 
     @Rule
     public final MockitoRule mockitoRule = MockitoJUnit.rule();
 
     @Mock
     private MoreSearch moreSearch;
+    @Mock
+    private Searches searches;
+    @Mock
+    private AggregationSearch.Factory aggregationSearchFactory;
+    @Mock
+    private AggregationSearch aggregationSearch;
+
+    private EventDefinitionDto eventDefinitionDto = EventDefinitionDto.builder()
+            .id("dto-id")
+            .title("Test Correlation")
+            .description("A test correlation event processors")
+            .config(getAggregationCountProcessorConfig())
+            .alert(false)
+            .keySpec(ImmutableList.of())
+            .notificationSettings(EventNotificationSettings.withGracePeriod(60000))
+            .priority(1)
+            .build();
 
     private AggregationCount subject;
+
+    private AggregationCountProcessorConfig getAggregationCountProcessorConfig() {
+        return AggregationCountProcessorConfig.builder()
+                .stream("main stream")
+                .thresholdType(ThresholdType.MORE.getDescription())
+                .threshold(threshold)
+                .searchWithinMs(2 * 60 * 1000)
+                .executeEveryMs(2 * 60 * 1000)
+                .groupingFields(new HashSet<>())
+                .distinctionFields(new HashSet<>())
+                .comment("test comment")
+                .searchQuery("*")
+                .build();
+    }
 
     @Test
     public void runCheckShouldPreciseMoreThanInTheResult() {
@@ -69,12 +106,12 @@ public class AggregationCountTest {
         int thresholdTest = 9;
 
         AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(type, thresholdTest, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
         final CountResult countResult = mock(CountResult.class);
         when(countResult.count()).thenReturn(thresholdTest + 1L);
-        when(moreSearch.count(anyString(), any(TimeRange.class), anyString())).thenReturn(countResult);
-        when(moreSearch.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
+        when(this.searches.count(anyString(), any(TimeRange.class), anyString())).thenReturn(countResult);
+        when(this.searches.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
 
         Result result = this.subject.runCheck(buildDummyTimeRange());
 
@@ -82,25 +119,27 @@ public class AggregationCountTest {
     }
 
     @Test
-    public void runCheckWithAggregateMorePositive() {
+    public void runCheckWithAggregateMorePositive() throws Exception {
         List<String> groupingFields = new ArrayList<>();
         groupingFields.add("user");
         groupingFields.add("ip_src");
         List<String> distinctionFields = new ArrayList<>();
-        AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(ThresholdType.MORE, threshold, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(ThresholdType.MORE, 2, groupingFields, distinctionFields);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
-        searchTermsThreeAggregateWillReturn(threshold + 1L);
-        when(moreSearch.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
+        mockFactorySearch();
+
+        searchTermsThreeAggregateWillReturn(threshold + 1);
+        when(this.searches.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
         Result result = this.subject.runCheck(buildDummyTimeRange());
 
-        String resultDescription = "Stream had 101 messages in the last 0 milliseconds with trigger condition more than 100 messages with the same value of the fields ip_src, user. (Executes every: 0 milliseconds)";
+        String resultDescription = "Stream had 3 messages in the last 0 milliseconds with trigger condition more than 2 messages with the same value of the fields ip_src, user. (Executes every: 0 milliseconds)";
         assertEquals(resultDescription, result.getResultDescription());
         assertEquals("Matching messages ", 3, result.getMessageSummaries().size());
     }
 
     @Test
-    public void runCheckWithAggregateLessPositive() {
+    public void runCheckWithAggregateLessPositive() throws Exception {
         final ThresholdType type = ThresholdType.LESS;
         List<String> groupingFields = new ArrayList<>();
         groupingFields.add("user");
@@ -108,46 +147,57 @@ public class AggregationCountTest {
         List<String> distinctionFields = new ArrayList<>();
         distinctionFields.add("user");
         AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(type, threshold, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
+        mockFactorySearch();
         searchTermsOneAggregateShouldReturn(threshold + 1L);
-        when(moreSearch.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
+        when(this.searches.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
 
         Result result = this.subject.runCheck(buildDummyTimeRange());
 
-        String resultDescription = "Stream had 1 messages in the last 0 milliseconds with trigger condition less than 100 messages with the same value of the fields " + String.join(", ", configuration.groupingFields())
+        String resultDescription = "Stream had 1 messages in the last 0 milliseconds with trigger condition less than " + threshold + " messages with the same value of the fields " + String.join(", ", configuration.groupingFields())
                 + ", and with distinct values of the fields " + String.join(", ", configuration.distinctionFields()) + ". (Executes every: 0 milliseconds)";
         assertEquals(resultDescription, result.getResultDescription());
         assertEquals("Matching messages ", 1, result.getMessageSummaries().size());
     }
 
     @Test
-    public void runCheckWithAggregateMoreNegative() {
+    public void runCheckWithAggregateMoreNegative() throws Exception {
         ThresholdType type = ThresholdType.MORE;
         List<String> groupingFields = new ArrayList<>();
         groupingFields.add("user");
         List<String> distinctionFields = new ArrayList<>();
         distinctionFields.add("user");
         AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(type, threshold, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
+        mockFactorySearch();
         searchTermsOneAggregateShouldReturn(threshold - 1L);
 
         Result result = this.subject.runCheck(buildDummyTimeRange());
         assertEquals("", result.getResultDescription());
         assertEquals("Matching messages ", 0, result.getMessageSummaries().size());
+
+
+// TODO
+//        AggregationSearch aggregationSearch = mock(AggregationSearch.class);
+//        when(this.aggregationSearchFactory.create(any(AggregationEventProcessorConfig.class),
+//                any(AggregationEventProcessorParameters.class ),
+//                anyString(),
+//                any(EventDefinition.class))).thenReturn(aggregationSearch);
     }
 
     @Test
-    public void runCheckWithAggregateLessNegative() {
+    public void runCheckWithAggregateLessNegative() throws Exception {
         ThresholdType type = ThresholdType.LESS;
         List<String> groupingFields = new ArrayList<>();
         groupingFields.add("user");
         List<String> distinctionFields = new ArrayList<>();
         AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(type, threshold, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
-        searchTermsThreeAggregateWillReturn(threshold +1L);
+        mockFactorySearch();
+        searchTermsThreeAggregateWillReturn(threshold + 1);
 
         Result result = this.subject.runCheck(buildDummyTimeRange());
         assertEquals("", result.getResultDescription());
@@ -155,20 +205,21 @@ public class AggregationCountTest {
     }
 
     @Test
-    public void runCheckWithAggregateMorePositiveWithNoBacklog() {
+    public void runCheckWithAggregateMorePositiveWithNoBacklog() throws Exception {
         ThresholdType type = ThresholdType.MORE;
         List<String> groupingFields = new ArrayList<>();
         groupingFields.add("user");
         groupingFields.add("ip_src");
         List<String> distinctionFields = new ArrayList<>();
         AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(type, threshold, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
-        searchTermsThreeAggregateWillReturn(threshold + 1L);
-        when(moreSearch.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
+        mockFactorySearch();
+        searchTermsThreeAggregateWillReturn(threshold + 1);
+        when(this.searches.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
 
         Result result = this.subject.runCheck(buildDummyTimeRange());
-        String resultDescription = "Stream had 101 messages in the last 0 milliseconds with trigger condition more than "
+        String resultDescription = "Stream had " + (threshold + 1) + " messages in the last 0 milliseconds with trigger condition more than "
                 + configuration.threshold() + " messages with the same value of the fields " + String.join(", ", configuration.groupingFields())
                 + ". (Executes every: 0 milliseconds)";
         assertEquals(resultDescription, result.getResultDescription());
@@ -183,12 +234,12 @@ public class AggregationCountTest {
         final int thresholdTest = 9;
 
         AggregationCountProcessorConfig configuration = getAggregationCountProcessorConfigWithFields(type, thresholdTest, groupingFields, distinctionFields);
-        this.subject = new AggregationCount(this.moreSearch, configuration);
+        this.subject = new AggregationCount(this.searches, this.moreSearch, configuration, this.eventDefinitionDto, this.aggregationSearchFactory);
 
         final CountResult countResult = mock(CountResult.class);
         when(countResult.count()).thenReturn(thresholdTest + 1L);
-        when(moreSearch.count(anyString(), any(TimeRange.class), anyString())).thenReturn(countResult);
-        when(moreSearch.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
+        when(this.searches.count(anyString(), any(TimeRange.class), anyString())).thenReturn(countResult);
+        when(this.searches.search(anyString(), anyString(), any(TimeRange.class), any(int.class), any(int.class), any(Sorting.class))).thenReturn(buildDummySearchResult());
 
         Result result = this.subject.runCheck(buildDummyTimeRange());
 
@@ -204,34 +255,70 @@ public class AggregationCountTest {
                 .threshold(threshold)
                 .searchWithinMs(0)
                 .executeEveryMs(0)
-                .groupingFields(new HashSet<>(groupingFields))
-                .distinctionFields(new HashSet<>(distinctionFields))
+                .groupingFields(new TreeSet<>(groupingFields))
+                .distinctionFields(new TreeSet<>(distinctionFields))
                 .comment("test comment")
                 .searchQuery("*")
                 .build();
     }
 
-    private void searchTermsOneAggregateShouldReturn(long count) {
-        final TermsResult termsResult = mock(TermsResult.class);
-        Map<String, Long> terms = new HashMap<String, Long>();
-        terms.put("user - ip1", count);
+    private void searchTermsOneAggregateShouldReturn(long count) throws Exception {
+        mockAggregationResult(1);
+//        final TermsResult termsResult = mock(TermsResult.class);
+//        Map<String, Long> terms = new HashMap<String, Long>();
+//        terms.put("user - ip1", count);
+//
+//        when(termsResult.terms()).thenReturn(terms);
 
-        when(termsResult.getTerms()).thenReturn(terms);
-
-        when(moreSearch.terms(anyString(), anyList() , any(int.class), anyString(), anyString(), any(TimeRange.class), any(Sorting.Direction.class))).thenReturn(termsResult);
+        // TODO Make test
+        //when(moreSearch.terms(anyString(), anyList(), any(int.class), anyString(), anyString(), any(TimeRange.class), any(Sorting.Direction.class))).thenReturn(termsResult);
 
     }
 
-    private void searchTermsThreeAggregateWillReturn(long count) {
-        final TermsResult termsResult = mock(TermsResult.class);
-        Map<String, Long> terms = new HashMap<String, Long>();
-        terms.put("user - ip1", count);
-        terms.put("user - ip2", count);
-        terms.put("user - ip3", count);
+    private void searchTermsThreeAggregateWillReturn(int count) throws Exception {
+        mockAggregationResult(count);
 
-        when(termsResult.getTerms()).thenReturn(terms);
+//        final TermsResult termsResult = mock(TermsResult.class);
+//        Map<String, Long> terms = new HashMap<String, Long>();
+//        terms.put("user - ip1", count);
+//        terms.put("user - ip2", count);
+//        terms.put("user - ip3", count);
+//
+//        when(termsResult.terms()).thenReturn(terms);
 
-        when(moreSearch.terms(anyString(), anyList() , any(int.class), anyString(), anyString(), any(TimeRange.class), any(Sorting.Direction.class))).thenReturn(termsResult);
+        // TODO Make test
+        //when(moreSearch.terms(anyString(), anyList(), any(int.class), anyString(), anyString(), any(TimeRange.class), any(Sorting.Direction.class))).thenReturn(termsResult);
+    }
+
+    private AggregationSeriesValue mockSeriesValue(int count, String... keys) {
+        AggregationSeriesValue aggregationSeriesValue = mock(AggregationSeriesValue.class);
+        when(aggregationSeriesValue.key()).thenReturn(ImmutableList.copyOf(keys));
+        when(aggregationSeriesValue.value()).thenReturn(Double.valueOf(count));
+        return aggregationSeriesValue;
+    }
+
+    private void mockFactorySearch() throws Exception {
+        when(this.aggregationSearchFactory.create(any(AggregationEventProcessorConfig.class),
+                any(AggregationEventProcessorParameters.class),
+                anyString(),
+                any(EventDefinition.class))).thenReturn(aggregationSearch);
+    }
+
+    private void mockAggregationResult(int count) throws Exception {
+        ImmutableList.Builder<AggregationSeriesValue> seriesValues = ImmutableList.builder();
+        for (int i = 1; i <= count; i++) {
+            AggregationSeriesValue aggregationSeriesValue = mockSeriesValue(count, "user", "ip" + i);
+            seriesValues.add(aggregationSeriesValue);
+        }
+
+        AggregationKeyResult aggregationKeyResult = mock(AggregationKeyResult.class);
+        when(aggregationKeyResult.seriesValues()).thenReturn(seriesValues.build());
+
+        AggregationResult aggregationResult = mock(AggregationResult.class);
+        when(aggregationResult.totalAggregatedMessages()).thenReturn(3L * count);
+        when(aggregationResult.keyResults()).thenReturn(ImmutableList.of(aggregationKeyResult));
+
+        when(aggregationSearch.doSearch()).thenReturn(aggregationResult);
     }
 
     private SearchResult buildDummySearchResult() {